+++ src/libs/conduit/python/conduit_python.cpp
@@ -1696,6 +1696,7 @@ PyConduit_DataType_to_string(PyConduit_DataType* self,
     return (Py_BuildValue("s", oss.str().c_str()));
 }
 
+
 //---------------------------------------------------------------------------//
 static PyObject *
 PyConduit_DataType_to_json(PyConduit_DataType* self,
@@ -4177,7 +4178,7 @@ PyConduit_Node_str(PyConduit_Node* self)
 static PyObject *
 PyConduit_Node_repr(PyConduit_Node* self)
 {
-    return PyConduit_Node_str(self);
+   return (Py_BuildValue("s", self->node->to_summary_string().c_str()));
 }
 
 //---------------------------------------------------------------------------//
@@ -5539,6 +5540,59 @@ PyConduit_Node_to_string(PyConduit_Node* self,
 
 //---------------------------------------------------------------------------//
 static PyObject *
+PyConduit_Node_to_summary_string(PyConduit_Node* self,
+                                 PyObject* args,
+                                 PyObject* kwargs)
+{
+    PyObject   *py_opts  = NULL;
+    static const char *kwlist[] = {"opts",
+                                   NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args,
+                                     kwargs,
+                                     "|O",
+                                     const_cast<char**>(kwlist),
+                                     &py_opts))
+    {
+        return (NULL);
+    }
+    
+    if(py_opts != NULL && !PyConduit_Node_Check(py_opts))
+    {
+        PyErr_SetString(PyExc_TypeError,
+                        "'opts' argument must be a "
+                        "conduit.Node instance");
+        return NULL;
+    }
+
+    Node opts;
+    Node *opts_ptr = &opts;
+    
+    if(py_opts != NULL)
+    {
+        opts_ptr = PyConduit_Node_Get_Node_Ptr(py_opts);
+    }
+
+    std::ostringstream oss;
+
+    try
+    {
+        self->node->to_summary_string_stream(oss,
+                                             *opts_ptr);
+
+    }
+    catch(conduit::Error e)
+    {
+        PyErr_SetString(PyExc_IOError,
+                        e.message().c_str());
+        return NULL;
+    }
+
+    return (Py_BuildValue("s", oss.str().c_str()));
+}
+
+//---------------------------------------------------------------------------//
+static PyObject *
 PyConduit_Node_to_json(PyConduit_Node* self,
                        PyObject* args,
                        PyObject* kwargs)
@@ -5959,6 +6013,12 @@ static PyMethodDef PyConduit_Node_METHODS[] = {
       "Returns a string representation of the node. "
       "Optionally takes protocol and spacing options. "
       "(Default protocol='yaml'.)"},
+     //------------------------------------------------------------t-----------//
+     {"to_summary_string",
+      (PyCFunction)PyConduit_Node_to_summary_string, 
+      METH_VARARGS| METH_KEYWORDS,
+      "Returns a summary string representation of the node. "
+      "Optionally takes a Node that provides spacing and threshold options. "},
      //-----------------------------------------------------------------------//
      {"to_json",
       (PyCFunction)PyConduit_Node_to_json,